Move K8sClient class from src/test into src/main
It will soon be used by ArmadaClientApplication.
