+++ src/test/scala/org/apache/spark/deploy/armada/e2e/ArmadaSparkE2E.scala
@@ -17,6 +17,8 @@
 
 package org.apache.spark.deploy.armada.e2e
 
+import org.apache.spark.deploy.armada.K8sClient
+
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
